LayoutTests/http/tests/security/mixedContent/insecure-xhr-sync-in-main-frame.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

</script>
<p>This test opens a HTTPS window that loads insecure data via synchronous XHR. We 
should trigger a mixed content callback because the main frame in the window is 
HTTPS but now has insecure data.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/insecure-xhr-sync-in-main-frame-window.html");
}
</script>
</body>
</html>
